# api/openapi/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthorBrief:
    properties:
      avatar:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
  dto.VideoInfo:
    properties:
      author:
        $ref: '#/definitions/dto.AuthorBrief'
      author_id:
        type: integer
      comment_count:
        type: integer
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      favorite_count:
        type: integer
      file_format:
        type: string
      file_size:
        type: integer
      height:
        type: integer
      id:
        type: integer
      play_url:
        type: string
      publish_time:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
      width:
        type: integer
    type: object
  dto.VideoListData:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
      videos:
        items:
          $ref: '#/definitions/dto.VideoInfo'
        type: array
    type: object
  dto.VideoUpdateRequest:
    properties:
      description:
        type: string
      status:
        enum:
        - pending
        - processing
        - published
        - failed
        - deleted
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
  response.ErrorInfo:
    properties:
      code:
        type: integer
      message:
        type: string
      type:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorInfo'
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: support@vida.com
    name: API Support
  description: 视频分享平台 API 服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vida-Go API
  version: "1.0"
paths:
...
  /videos/{id}:
    delete:
      description: 删除指定的视频
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 视频不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除视频
      tags:
      - 视频
    get:
      description: 根据视频ID获取视频详细信息
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.VideoInfo'
              type: object
        "404":
          description: 视频不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取视频详情
      tags:
      - 视频
    put:
      consumes:
      - application/json
      description: 更新视频的标题、描述等信息
      parameters:
      - description: 视频ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VideoUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.VideoInfo'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 视频不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新视频信息
      tags:
      - 视频
  /videos/feed:
    get:
      description: 获取视频列表（公开接口，不需要登录）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.VideoListData'
              type: object
      summary: 获取视频流
      tags:
      - 视频
  /videos/my/list:
    get:
      description: 获取当前用户上传的视频列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 视频状态筛选
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.VideoListData'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取我的视频列表
      tags:
      - 视频
  /videos/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传视频文件，支持 mp4, avi, mov, mkv, flv, webm 格式
      parameters:
      - description: 视频标题
        in: formData
        name: title
        required: true
        type: string
      - description: 视频描述
        in: formData
        name: description
        type: string
      - description: 视频文件
        in: formData
        name: video_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 上传视频
      tags:
      - 视频
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"